CLEANUP: code organisation and naming
